requirements.txt
streamlit
python-dotenv
supabase
postgrest
storage3

python-docx
openpyxl
pandas
numpy

google-generativeai
requests